move tests in root
