--- pb-rs/run_test.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-
-cargo run ../examples/codegen/data_types_import.proto
-cargo run ../examples/codegen/data_types.proto
-cd ..
-cargo run --example codegen_example
-
-cd tests/rust_protobuf
-./generate.sh
