--- pb-rs/test.sh
@@ -1,7 +1,0 @@
-cd ../tests/rust_protobuf
-./generate.sh
-
-cd ../issue69
-./generate.sh
-
-cargo test
