+++ quick-protobuf/examples/pb_rs/a/b.rs
