+++ quick-protobuf/examples/pb_rs/generate_module.sh
